音が一回だけなったときのコード
