+++ src/App.jsx
@@ -1,7 +1,6 @@
 import "./App.css"
 import { Grid } from './Grid'
 import { Header } from './Header'
-import { usePlay, useStop } from "./hooks/useSoundPlayer";
 
 export const App = () => {
