+++ src/Header.jsx
@@ -1,5 +1,4 @@
 import { DisplayCard } from "./DisplayCard";
-import { usePlay, useStop } from "./hooks/useSoundPlayer";
 
 export const Header = () =>{
     const styleHeader = {
@@ -30,30 +29,23 @@ export const Header = () =>{
             if(DisplayCards[i].dataset.occupied == "false"){ //カードが無かったら処理を停止
                 setTimeout( () =>{
                     DisplayCards[i-1].style.backgroundColor = "#3282B8"
-                    useStop(DisplayCard[i-1].innerHTML);
                 }, i * 1000) //選んだカードの一番後ろのオレンジ灯火を戻すための処理
                 break;
             }
             if(i  == 0){ //一個目はすぐに色を変える
                 DisplayCards[i].style.backgroundColor = "orange";
-                usePlay(DisplayCard[i].innerHTML);
             }else if(i < 3){ //初回、最終回以外の処理はここ
                 setTimeout( () =>{
                     DisplayCards[i-1].style.backgroundColor = "#3282B8";
-                    useStop(DisplayCard[i-1].innerHTML);
                     DisplayCards[i].style.backgroundColor = "orange";
-                    usePlay(DisplayCard[i].innerHTML);
                 }, i * 1000); //二つ目は１秒、三つ目は２秒待つ... とすることで１秒ごと動作させる
             } else if(i == 3){
                 setTimeout( () =>{
                     DisplayCards[i-1].style.backgroundColor = "#3282B8";
-                    useStop(DisplayCard[i-1].innerHTML);
                     DisplayCards[i].style.backgroundColor = "orange";
-                    usePlay(DisplayCard[i].innerHTML);
                 }, i * 1000); //二つ目は１秒、三つ目は２秒待つ... とすることで１秒ごと動作させる
                 setTimeout( () =>{
                     DisplayCards[i].style.backgroundColor = "#3282B8";  
-                    useStop(DisplayCard[i].innerHTML);  
                 }, (i+1) * 1000);
             }
